// src/app/enums/Feelings.ts
export enum Feelings {
  HAPPY = 'happy',
  POWERFUL = 'powerful',
  MOTIVATED = 'motivated',
  PROUD = 'proud',
  RELAXED = 'relaxed',
  SATISFIELD = 'satisfield',
  GRATEFUL = 'grateful',
  CALM = 'calm',
  LOVED = 'loved',
  FINE = 'fine',
  SAD = 'sad',
  LAZY = 'lazy',
  FRUSTRATED = 'frustrated',
  BORED = 'bored',
  ANXIOUS = 'anxious',
  ANNOYED = 'annoyed',
  TIRED = 'tired',
  DEPRESSED = 'depressed',
  LONELY = 'lonely',
  SICK = 'sick',
}